requirements.txt
streamlit
sentence-transformers
faiss-cpu
datasets
numpy
moviepy
librosa
soundfile
torch
openai-whisper
soundfile
ffmpeg-python
